requirements.txt
Django==1.8.2
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
kombu==3.0.30
pytz==2015.4
redis==2.10.3
requests==2.7.0
psycopg2-binary==2.8.6
djangorestframework==3.6.4
django-cors-headers==2.4.1
